* fxied compiler warning

+++ src/common/string.c
@@ -54,7 +54,7 @@ char *strscpy(char *dest, const char *src, size_t size)
 {
        register size_t l = 0;
 #if !defined(NODEBUG) && defined(DEBUG)
-       if (sizeof(dest) == size) fprintf(stderr, "Warning: probably sizeof(pointer) in strscpy(%p, %s, %d)!\n", dest, src, size);
+       if (sizeof(dest) == size) fprintf(stderr, "Warning: probably sizeof(pointer) in strscpy(%p, %s, %ld)!\n", dest, src, (long)size);
        if (NULL == dest) fprintf(stderr, "Warning: dest is NULL in strscpy!\n");
        if (NULL == src) fprintf(stderr, "Warning: src is NULL in strscpy!\n");
 #endif
